refactor wire package to satisfy file limit
