--- pkg/wire/generate.go
@@ -1,3 +1,0 @@
-package wire
-
-//go:generate go run github.com/google/wire/cmd/wire
